pyproject.toml
[tool.poetry]
name = "nflMatchupPredictor"
version = "0.1.0"
description = "NFL Matchup Predictor"
authors = ["Jonathan Armitage <jarmitage@developertown.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.0"
numpy = "^1.21.0"
click = "^8.0.3"
beautifulsoup4 = "^4.10.0"
html5lib = "^1.1"
lxml = "^4.7.1"
spyder = "^5.2.2"
scikit-learn = "^1.0.2"

[tool.poetry.dev-dependencies]
matplotlib = "^3.5.1"
seaborn = "^0.11.2"
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pre-commit = "^2.17.0"
pydocstyle = "^6.1.1"
Sphinx = "^4.4.0"
sphinx-rtd-theme = "^1.0.0"
python-dotenv = "^0.19.2"
notebook = "^6.4.8"
jupyter-contrib-nbextensions = "^0.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127
target-version = ["py39"]
extend-exclude = "nflMatchupPredictor/Features/Features.py"
